+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 ## [0.0.1] - [Unreleased] - [Billy Purvis] - 05-01-2017 
 
 ### Added 
+- Slick.css cdn
 - Added VueRouter
 - Added Foo.vue component
 - Added router-view/router-link to App.vue
@@ -15,6 +16,7 @@ All notable changes to this project will be documented in this file.
 
 ### Changed
 - added 'created' lifecycle hook to request methods automatically
+- changed 'created' hook to 'mounted' to wait until DOM is ready
 
 ### Removed
 - Slick content